@@ -25,32 +25,32 @@
   },
   "devDependencies": {
     "@electron-forge/cli": "^7.2.0",
-    "@vitejs/plugin-vue": "^4.4.1",
+    "@vitejs/plugin-vue": "^4.5.2",
     "autoprefixer": "^10.4.16",
-    "electron": "^27.0.4",
-    "electron-builder": "^24.6.4",
-    "postcss": "^8.4.31",
-    "tailwindcss": "^3.3.5",
-    "typescript": "^5.2.2",
-    "unplugin-auto-import": "^0.17.1",
+    "electron": "^27.2.0",
+    "electron-builder": "^24.9.1",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.4.0",
+    "typescript": "^5.3.3",
+    "unplugin-auto-import": "^0.17.2",
     "unplugin-vue-components": "^0.25.2",
-    "vite": "^5.0.4",
-    "vite-plugin-electron": "^0.15.4",
+    "vite": "^5.0.10",
+    "vite-plugin-electron": "^0.15.5",
     "vite-plugin-electron-renderer": "^0.14.5",
-    "vue": "^3.3.8",
+    "vue": "^3.3.13",
     "vue-router": "^4.2.5",
-    "vue-tsc": "^1.8.22"
+    "vue-tsc": "^1.8.26"
   },
   "dependencies": {
-    "@element-plus/icons-vue": "^2.1.0",
-    "@octokit/core": "^5.0.1",
-    "axios": "^1.6.1",
+    "@element-plus/icons-vue": "^2.3.1",
+    "@octokit/core": "^5.0.2",
+    "axios": "^1.6.2",
     "electron-store": "^8.1.0",
-    "element-plus": "^2.4.2",
-    "enkanetwork.js": "^2.5.3",
-    "marked": "^11.0.0",
+    "element-plus": "^2.4.4",
+    "enkanetwork.js": "^2.6.2",
+    "marked": "^11.1.0",
     "path": "^0.12.7",
     "sass": "^1.69.5",
-    "vue-i18n": "^9.6.5"
+    "vue-i18n": "^9.8.0"
   }
 }
