@@ -1,4 +1,4 @@
-<script setup>
+<script setup lang="ts">
 const props = defineProps({
     banners: {}
 })
@@ -10,8 +10,8 @@ const openLink = (url) => {
 
 <template>
     <el-carousel arrow="hover" :interval="5000" indicator-position="none">
-        <el-carousel-item class=" cursor-pointer" v-for="banner in banners" @click="openLink(banner.url)">
-            <img class=" object-scale-down" :src="banner.img" />
+        <el-carousel-item class="cursor-pointer" v-for="banner in banners" @click="openLink(banner.url)">
+            <img class="object-scale-down" :src="banner.img" />
         </el-carousel-item>
     </el-carousel>
 </template>
