@@ -1,7 +1,7 @@
-<script setup>
+<script setup lang="ts">
 import { onMounted, ref, computed } from 'vue'
 import { useRouter } from 'vue-router'
-import { translate } from '../i18n/index'
+import { translate } from '../i18n'
 const gameName = translate('general_hi3')
 
 const hiLauncherPath = ref('')
@@ -18,7 +18,7 @@ const launcherInfoReady = ref(false)
 const launcherInfoFailed = ref(false)
 const errMsg = ref('')
 const hideElements = ref(false)
-const scrollbarref = ref()
+const scrollBarRef = ref()
 const importDialogShow = ref(false)
 const combinePaths = ref(true)
 
@@ -254,7 +254,7 @@ const onImportDialogClose = () => {
             class="absolute left-16 top-96 z-50 rounded-xl transition-all backdrop-blur-md pl-3 pr-1 font-bold"
             :class="hideElements ? 'opacity-0 -translate-y-2 pointer-events-none blur-md -translate-x-14 scale-110' : 'opacity-100 pointer-events-auto'"
             style="height: 125px; width: 396px; background-color: rgb(255 255 255 / 0.7); transition-duration: 500ms;" />
-        <el-scrollbar ref="scrollbarref" height="91vh" class="scroll-wrapper absolute z-40" @scroll="handleScroll">
+        <el-scrollbar ref="scrollBarRef" height="91vh" class="scroll-wrapper absolute z-40" @scroll="handleScroll">
             <div class="items-scroll flex flex-col content-center items-center w-full">
                 <div class="w-full flex flex-row justify-between">
                     <div class="w-1"></div>
