@@ -1,7 +1,7 @@
-<script setup>
-import { ref, onMounted } from 'vue'
+<script setup lang="ts">
+import {ref, onMounted, computed} from 'vue'
 import { Picture, RefreshLeft } from '@element-plus/icons-vue'
-import { translate } from '../i18n/index'
+import { translate } from '../i18n'
 
 const gsGamePath = ref('')
 const srGamePath = ref('')
@@ -10,6 +10,9 @@ const transitionShow = ref(false)
 const bgPath = ref('')
 
 const DEFAULT_BG = '../../src/assets/gsbanner.png'
+const bgImage = computed(() => {
+  return bgPath.value ? bgPath.value : DEFAULT_BG
+})
 
 onMounted(async () => {
     // For electron-store test
@@ -19,8 +22,6 @@ onMounted(async () => {
     srGamePath.value = await window.store.get('srGamePath')
     hi3GamePath.value = await window.store.get('hi3GamePath')
     bgPath.value = await window.store.get('mainBgPath')
-    const imgElement = document.getElementById('bgImage');
-    imgElement.src = bgPath.value ? bgPath.value : DEFAULT_BG;
     transitionShow.value = true
 })
 
@@ -28,7 +29,7 @@ const genshin = async () => {
     await window.child.exec(gsGamePath.value)
     const trayOnLaunch = await window.store.get('trayOnLaunch')
     if (trayOnLaunch) {
-        window.win.tray()
+        await window.win.tray()
     }
 }
 
@@ -36,7 +37,7 @@ const starRail = async () => {
     await window.child.exec(srGamePath.value)
     const trayOnLaunch = await window.store.get('trayOnLaunch')
     if (trayOnLaunch) {
-        window.win.tray()
+        await window.win.tray()
     }
 }
 
@@ -44,7 +45,7 @@ const honkai3 = async () => {
     await window.child.exec(hiGamePath.value)
     const trayOnLaunch = await window.store.get('trayOnLaunch')
     if (trayOnLaunch) {
-        window.win.tray()
+        await window.win.tray()
     }
 }
 
@@ -57,12 +58,9 @@ const setPic = async () => {
         if (resp) {
             bgPath.value = resp
             window.store.set('mainBgPath', bgPath.value, false)
-
-            const imgElement = document.getElementById('bgImage');
-            imgElement.src = resp;
         }
     }).catch((error) => {
-        console.error('Error in showing dialog:', error);
+        console.error('Error in showing dialog:', error)
     });
 }
 
@@ -75,20 +73,20 @@ const resetPic = () => {
 <template>
     <div @touchmove.prevent @mousewheel.prevent :class="transitionShow ? '' : 'opacity-0 blur-lg scale-90'"
         style="transition-duration: 400ms;">
-        <img class="bg-pic object-cover" id="bgImage" :src="bgPath ? bgPath : '../../src/assets/gsbanner.png'" />
+        <img class="bg-pic object-cover" :src="bgPath ? bgPath : '../../src/assets/gsbanner.png'" alt="Background image of Home page"/>
         <div class="sticky bottom-0" style="height: 60vh;"></div>
         <div class="bottom-area sticky">
             <h1 class="font-sans font-bold text-5xl" style="margin-bottom: 10px; font-family: 'genshin-font';">{{
-                $t('mainpage_title') }}</h1>
+                translate('mainpage_title') }}</h1>
             <button v-if="gsGamePath" @click="genshin"
                 class="p-3 mx-2 my-2 font-bold text-xl transition-all bg-yellow-400 hover:bg-yellow-500 active:bg-yellow-800 active:scale-90 cursor-default">{{
-                    $t('mainpage_buttonText', { game: $t('general_gsShort') }) }}</button>
+                    translate('mainpage_buttonText', { game: $t('general_gsShort') }) }}</button>
             <button v-if="srGamePath" @click="starRail"
                 class="p-3 mx-2 my-2 font-bold text-xl transition-all bg-yellow-400 hover:bg-yellow-500 active:bg-yellow-800 active:scale-90 cursor-default">{{
-                    $t('mainpage_buttonText', { game: $t('general_srShort') }) }}</button>
+                    translate('mainpage_buttonText', { game: $t('general_srShort') }) }}</button>
             <button v-if="hi3GamePath" @click="honkai3"
                 class="p-3 mx-2 my-2 font-bold text-xl transition-all bg-yellow-400 hover:bg-yellow-500 active:bg-yellow-800 active:scale-90 cursor-default">{{
-                    $t('mainpage_buttonText', { game: $t('general_hi3Short') }) }}</button>
+                    translate('mainpage_buttonText', { game: $t('general_hi3Short') }) }}</button>
         </div>
         <div class="rounded-full fixed left-8 bottom-4 w-10 h-10 bg-white hover:bg-gray-100 active:bg-gray-400 active:scale-90 transition-all"
             @click="setPic">
